+++ samplesheets/views.py
@@ -104,9 +104,9 @@ MISC_FILES_COLL = 'MiscFiles'
 TRACK_HUBS_COLL = 'TrackHubs'
 RESULTS_COLL_ID = 'results_reports'
 RESULTS_COLL = 'ResultsReports'
-IRODS_REQ_CREATE_ALERT_NAME = 'irods_request_create'
-IRODS_REQ_ACCEPT_ALERT_NAME = 'irods_request_accept'
-IRODS_REQ_REJECT_ALERT_NAME = 'irods_request_reject'
+IRODS_REQ_CREATE_ALERT = 'irods_request_create'
+IRODS_REQ_ACCEPT_ALERT = 'irods_request_accept'
+IRODS_REQ_REJECT_ALERT = 'irods_request_reject'
 SYNC_SUCCESS_MSG = 'Sample sheet sync successful'
 SYNC_FAIL_DISABLED = 'Sample sheet sync disabled'
 SYNC_FAIL_PREFIX = 'Sample sheet sync failed'
@@ -780,8 +780,7 @@ class IrodsRequestModifyMixin:
 
     # App Alert Helpers --------------------------------------------------------
 
-    @classmethod
-    def add_alerts_create(cls, project, app_alerts=None):
+    def add_alerts_create(self, project, app_alerts=None):
         """
         Add app alerts for project owners/delegates on request creation. Will
         not create new alerts if the user already has a similar active alert
@@ -803,10 +802,12 @@ class IrodsRequestModifyMixin:
         )
         # logger.debug('od_users={}'.format(od_users))  # DEBUG
         for u in od_users:
+            if u == self.request.user:
+                continue  # Skip triggering user
             alert_count = AppAlert.objects.filter(
                 project=project,
                 user=u,
-                alert_name=IRODS_REQ_CREATE_ALERT_NAME,
+                alert_name=IRODS_REQ_CREATE_ALERT,
                 active=True,
             ).count()
             if alert_count > 0:
@@ -814,7 +815,7 @@ class IrodsRequestModifyMixin:
                 continue  # Only have one active alert per user/project
             app_alerts.add_alert(
                 app_name=APP_NAME,
-                alert_name=IRODS_REQ_CREATE_ALERT_NAME,
+                alert_name=IRODS_REQ_CREATE_ALERT,
                 user=u,
                 message='iRODS delete requests require attention in '
                 'project "{}"'.format(project.title),
@@ -851,7 +852,7 @@ class IrodsRequestModifyMixin:
         )
         if req_count == 0:
             alerts = AppAlert.objects.filter(
-                alert_name=IRODS_REQ_CREATE_ALERT_NAME,
+                alert_name=IRODS_REQ_CREATE_ALERT,
                 project=irods_request.project,
                 active=True,
             )
@@ -2344,7 +2345,7 @@ class IrodsRequestAcceptView(
             )
 
         # Prepare and send notification email
-        if settings.PROJECTROLES_SEND_EMAIL:
+        if settings.PROJECTROLES_SEND_EMAIL and obj.user != self.request.user:
             subject_body = 'iRODS delete request accepted'
             message_body = EMAIL_DELETE_REQUEST_ACCEPT.format(
                 project=obj.project.title,
@@ -2357,10 +2358,10 @@ class IrodsRequestAcceptView(
             )
 
         # Create app alert
-        if app_alerts:
+        if app_alerts and obj.user != self.request.user:
             app_alerts.add_alert(
                 app_name=APP_NAME,
-                alert_name=IRODS_REQ_ACCEPT_ALERT_NAME,
+                alert_name=IRODS_REQ_ACCEPT_ALERT,
                 user=obj.user,
                 message='iRODS delete request accepted by {}: "{}"'.format(
                     self.request.user.username, obj.get_short_path()
@@ -2437,7 +2438,7 @@ class IrodsRequestRejectView(
         )
 
         # Prepare and send notification email
-        if settings.PROJECTROLES_SEND_EMAIL:
+        if settings.PROJECTROLES_SEND_EMAIL and obj.user != self.request.user:
             subject_body = 'iRODS delete request rejected'
             message_body = EMAIL_DELETE_REQUEST_REJECT.format(
                 project=obj.project.title,
@@ -2450,10 +2451,10 @@ class IrodsRequestRejectView(
             )
 
         # Create app alert
-        if app_alerts:
+        if app_alerts and obj.user != self.request.user:
             app_alerts.add_alert(
                 app_name=APP_NAME,
-                alert_name=IRODS_REQ_REJECT_ALERT_NAME,
+                alert_name=IRODS_REQ_REJECT_ALERT,
                 user=obj.user,
                 message='iRODS delete request rejected by {}: "{}"'.format(
                     self.request.user.username, obj.get_short_path()