+++ taskflowbackend/flows/landing_zone_delete.py
@@ -66,4 +66,3 @@ class Flow(BaseLinearFlow):
                 },
             )
         )
-